Parse natural language dates
Currently you have to input a date in a survey precisely matching
the Django INPUT_FORMATS.

This commit uses dateparser to process a date which somebody enters
as natural language.

It looks like it performs adequate sanitisation of user input.

The parsing is quite expensive and the Django docs recommend to
implement caching on value_from_datadict() because it can be
called multiple times in one request, so I've done that.